Split codegen for return into module
